% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cbind}
\alias{cbind}
\title{A cbind variation that ignores objects with zero dimention}
\usage{
cbind(...)
}
\arguments{
\item{...}{(generalized) vectors or matrices. These can be
given as named arguments. Other R objects may be coerced as
appropriate, or S4 methods may be used: see sections ‘Details’
and ‘Value’. (For the "data.frame" method of cbind these can
be further arguments to data.frame such as stringsAsFactors.)}
}
\description{
A cbind variation that ignores objects with zero dimention
}
