% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBalancedFolds.R
\name{CreateBalancedFolds}
\alias{CreateBalancedFolds}
\title{Utility function to create cross-validation with each fold balanced}
\usage{
CreateBalancedFolds(y, k = 10, list = TRUE, returnTrain = FALSE)
}
\arguments{
\item{y}{is a data.frame #add more Gustavo?}

\item{k}{is the number of partiions (default is 10)}
}
\description{
This function allows your dataframe to be be partitioned for multi-fold cross-validation with a balanced number of classes (e.g.same number of presences absences). Called as part of the mcaret function.
}
\keyword{cross-validation}
