% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBalancedMultiFolds.R
\name{CreateBalancedMultiFolds}
\alias{CreateBalancedMultiFolds}
\title{Utility function to create multi-fold cross-validation with each fold balanced}
\usage{
CreateBalancedMultiFolds(y, k = 10, times = 5)
}
\arguments{
\item{y}{is a data.frame}

\item{k}{is the number of partiions, default is 10}

\item{times}{is the number of times the data is partioned , default is 5.}
}
\description{
This function allows your dataframe to be be partitioned for multi-fold cross-validation with a balanced number of classes (e.g.same number of presences absences). Called as part of the mcaret function.
}
\keyword{cross-validation}
