% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devianceResids.R
\name{devianceResids}
\alias{devianceResids}
\title{Calculate binary outcome deviance residuals using probability predictions}
\usage{
devianceResids(yhat, truth)
}
\arguments{
\item{yhat}{A \code{vector} of predictions on the probability scale.}

\item{truth}{A \code{vector} of true binary observations in the same order as \code{yhat}}
}
\description{
Calculate binary outcome deviance residuals using probability predictions
}
\details{
This function uses the predictions from a model applied to binary outcomes to
calculate deviance residuals. Predictions must be on the probability scale.
}
