% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Features.R
\docType{data}
\name{Features}
\alias{Features}
\title{Landscape and host feature (variable) data}
\format{

}
\source{
Fountain‐Jones, NM, Craft, ME, Funk, WC, et al. Urban landscapes can change virus gene flow and evolution in a fragmentation‐sensitive carnivore. Mol Ecol. 2017; 26: 6487– 6498. https://doi.org/10.1111/mec.14375
}
\usage{
Features
}
\description{
Landscape and host feature (variable) data
}
\examples{
\dontrun{
 Features
}
}
\keyword{datasets}
