% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrCovar.R
\name{mr_Covar}
\alias{mr_Covar}
\title{Investigate partial dependencies of a covariate for mrIML JSDMs (Joint
species distirbution models)}
\usage{
mr_Covar(mrIMLobj, var, sdthresh = 0.05)
}
\arguments{
\item{mrIMLobj}{A list object output by \code{\link[=mrIMLpredicts]{mrIMLpredicts()}}.}

\item{var}{The variable of interest for calculating the profile.}

\item{sdthresh}{The standard deviation threshold for filtering taxa
(default: 0.05).}
}
\value{
A plot displaying the covariate partial dependence profiles for those
models that meet the \code{sdthreshold} requirement and another summarizing the
rates of change in probability for the specified variable
(the derivatives of the PD curves).
}
\description{
This function is a wrapper around \code{\link[=mrFlashlight]{mrFlashlight()}} that plots the covariate
partial dependencies for a specified environmental/host variable. It also
filters the taxa based on standard deviation thresholds.
}
\examples{
library(tidymodels)

# Without bootstrap
data <- MRFcov::Bird.parasites
Y <- data \%>\%
  select(-scale.prop.zos) \%>\%
  select(order(everything()))
X <- data \%>\%
  select(scale.prop.zos)

model_rf <- rand_forest(
  trees = 100, # 100 trees are set for brevity. Aim to start with 1000
  mode = "classification",
  mtry = tune(),
  min_n = tune()
) \%>\%
  set_engine("randomForest")

mrIML_rf <- mrIMLpredicts(
  X = X,
  Y = Y,
  X1 = Y,
  Model = model_rf,
  prop = 0.7,
  k = 5
)

mrIML_rf \%>\%
  mr_Covar(var = "scale.prop.zos", sdthresh = 0.05)

}
