% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrBenchmark.R
\name{mrBenchmark}
\alias{mrBenchmark}
\title{Compare and benchmark disease outbreak risk among and within groups}
\usage{
mrBenchmark(
  my.data = "data",
  outcome = "class",
  y = "predicted",
  group = "group1",
  label_by = "ID",
  type = "internal"
)
}
\arguments{
\item{my.data}{A \code{character} object name of data frame}

\item{outcome}{A \code{character} column name of variable containing outcome}

\item{y}{A \code{character} column name of variable containing model predicted values}

\item{group}{A \code{character} column name of variable that individuals should be grouped by}

\item{label_by}{A \code{character} column name of variable representing the individual units. If stated, these will be labeled on the ggplot. By default labels will not be included}

\item{type}{A \code{character} specify within group "internal" or among group "external" benchmarking}
}
\value{

}
\description{
Compare and benchmark disease outbreak risk among and within groups
}
