% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrBootstrap.R
\name{mrBootstrap}
\alias{mrBootstrap}
\title{Bootstrap mrIML model predictions}
\usage{
mrBootstrap(mrIMLobj, num_bootstrap = 10, downsample = FALSE)
}
\arguments{
\item{mrIMLobj}{A list object output by \code{\link[=mrIMLpredicts]{mrIMLpredicts()}}.}

\item{num_bootstrap}{The number of bootstrap samples to generate
(default: 10).}

\item{downsample}{Logical. Should the bootstrap samples be downsampled?
(default: FALSE).}
}
\value{
A list containing bootstrap samples of variable profiles for each
response variable.
}
\description{
This function bootstraps model predictions and generates variable profiles
for each response variable.
}
\examples{
!identical(Sys.getenv("NOT_CRAN"), "true")
library(tidymodels)

data <- MRFcov::Bird.parasites
Y <- data \%>\%
  select(-scale.prop.zos) \%>\%
  select(order(everything()))
X <- data \%>\%
  select(scale.prop.zos)

# Specify a random forest tidy model
model_rf <- rand_forest(
  trees = 100, # 100 trees are set for brevity. Aim to start with 1000
  mode = "classification",
  mtry = tune(),
  min_n = tune()
) \%>\%
  set_engine("randomForest")

mrIML_rf <- mrIMLpredicts(
  X = X,
  Y = Y,
  X1 = Y,
  Model = model_rf,
  prop = 0.7,
  k = 5
)

# cl <- parallel::makeCluster(5)
# future::plan(cluster, workers=cl)

mrIML_bootstrap <- mrIML_rf \%>\%
  mrBootstrap(
  num_bootstrap = 50
  )
}
