% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrIML-package.R
\docType{package}
\name{mrIML-package}
\alias{mrIML}
\alias{mrIML-package}
\title{mrIML: Multi Response Interpretable Machine Learning}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package aims to enable users to build and interpret multivariate machine learning models harnessing the tidyverse (tidy model syntax in particular). This package builds off ideas from Gradient Forests (Ellis et al 2012), ecological genomic approaches (Fitzpatrick and Keller, 2014) and multi-response stacking algorithms (Xing et al 2019).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nfj1380/mrIML}
  \item Report bugs at \url{https://github.com/nfj1380/mrIML/issues}
}

}
\author{
\strong{Maintainer}: Nick Fountain-Jones \email{nfountainjones@gmail.com} (\href{https://orcid.org/0000-0001-9248-8493}{ORCID})

Authors:
\itemize{
  \item Gustavo Machado \email{gmachad@ncsu.edu} (\href{https://orcid.org/0000-0001-7552-6144}{ORCID})
  \item Chris Kozakiewicz
  \item Nick Clark
}

}
\keyword{internal}
