% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrIML_bird_parasites_RF.R
\docType{data}
\name{mrIML_bird_parasites_RF}
\alias{mrIML_bird_parasites_RF}
\title{An example mrIML model fit to \link[MRFcov:Bird.parasites]{MRFcov::Bird.parasites}}
\format{
An object of class \code{list} of length 3.
}
\usage{
mrIML_bird_parasites_RF
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{data <- MRFcov::Bird.parasites
Y <- data \%>\%
  dplyr::select(-scale.prop.zos) \%>\%
  dplyr::select(order(everything()))
X <- data \%>\%
  dplyr::select(scale.prop.zos)
}\if{html}{\out{</div>}}
}
\details{
model_rf <- parsnip::rand_forest(
trees = 10, # 10 trees are set for brevity. Aim to start with 1000
mode = "classification",
mtry = tune::tune(),
min_n = tune::tune(),
engine = "randomForest"
)

mrIML_bird_parasites <- mrIMLpredicts(
X = X,
Y = Y,
X1 = Y,
Model = model_rf,
prop = 0.7,
k = 2,
racing = FALSE
)

\if{html}{\out{<div class="sourceCode">}}\preformatted{}\if{html}{\out{</div>}}
}
\keyword{datasets}
