% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MrIMLconverts.R
\name{MrIMLconverts}
\alias{MrIMLconverts}
\title{Converts MrIML objects into a format suitable for IML R package}
\usage{
MrIMLconverts(yhats, X, mode = "regression")
}
\arguments{
\item{yhats}{A \code{list} is the list generated by mrIMLpredicts}

\item{mode}{\code{character}'classification' or 'regression' i.e., is the generative model a regression or classification?}
}
\description{
Converts MrIML objects into a format suitable for IML R package
}
\details{
This function is a wrapper to convert MrIML objects into the IML R package format (Molnar and Schratz, 2021)
}
