% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrIMLperformance.R
\name{mrIMLperformance}
\alias{mrIMLperformance}
\title{Calculate general performance metrics of a mrIML model}
\usage{
mrIMLperformance(mrIMLobj)
}
\arguments{
\item{mrIMLobj}{A list object created by \code{\link[=mrIMLpredicts]{mrIMLpredicts()}} containing
multi-response models.}
}
\value{
A list with two slots.
\itemize{
\item \verb{$model_performance}: A \pkg{tibble} of commonly used metric that can be used
to compare model performance of classification models. Performance metrics
are based on the test data defined during \code{\link[=mrIMLpredicts]{mrIMLpredicts()}}.
\item \verb{$global_performance_summary}: A global performance metric: the average of
a performance metric over all response models. MMC is used for classification
models and RMSE for regression model.
}
}
\description{
Summarizes the performance of a \code{mrIML} object created using
\code{\link[=mrIMLpredicts]{mrIMLpredicts()}} in a way so that different models can be easily compared.
For regression models, root mean squared error (RMSE) and R-squared are
reported, while for classification models, area under the ROC (AUC), Mathews
correlation coefficient (MMC), positive predictive value (PPC), specificity,
and sensitivity are reported.
}
\examples{
library(tidymodels)

data <- MRFcov::Bird.parasites

# Prepare data for mrIML
Y <- data \%>\%
  select(-scale.prop.zos) \%>\%
  select(order(everything()))
X <- data \%>\%
  select(scale.prop.zos)
X1 <- Y

# Fit GN model
model_rf <- rand_forest(
  trees = 100, # 100 trees are set for brevity. Aim to start with 1000
  mode = "classification",
  mtry = tune(),
  min_n = tune()
) \%>\%
  set_engine("randomForest")

GN_model_rf <- mrIMLpredicts(
  X = X,
  Y = Y,
  X1 = X1,
  Model = model_rf,
  prop = 0.7,
  k = 5
)

perf <- mrIMLperformance(GN_model_rf)
perf[[1]]
perf[[2]]

}
