% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{mrLocalExplainer}
\alias{mrLocalExplainer}
\title{Run local explanation methods for individual data points}
\usage{
mrLocalExplainer(X, Model, Y)
}
\arguments{
\item{X}{A \code{dataframe} data frame of predictor variables}

\item{Model}{A \code{workflow} workflow object containing the machine learning model}

\item{Y}{\code{vector} vector containing the outcome for each data instance}
}
\description{
Run local explanation methods for individual data points
}
