% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MrProfile2D.R
\name{mrProfile2D}
\alias{mrProfile2D}
\title{Plots multi-response 2D profile plots to help interpret interactions}
\usage{
mrProfile2D(
  mrIMLconverts_list,
  featureA,
  featureB,
  mode = "regression",
  grid.size = 30,
  method = "ale"
)
}
\arguments{
\item{mrIMLconverts_list}{A \code{list} list of iml objects generated by MrIMLconverts}

\item{featureA}{A \code{character} the name of the first feature the interaction pair}

\item{featureB}{A \code{character} the name of the second feature the interaction pair}

\item{mode}{\code{character}'classification' or 'regression' i.e., is the generative model a regression or classification?}

\item{grid.size}{\code{numeric} number of grid points to define the resolution of partial dependencies or ALEs across feature space. .}

\item{method}{\code{character} method used to construct the 2D plot. Can be either 'ale' (default) or 'pdp'}
}
\description{
Plots multi-response 2D profile plots to help interpret interactions
}
\details{
This function computes multivariate 2D profile plots for feature pairs of interest by taking the average prediction across feature space for all responses.
This plot can be generated based on accumulated feature effects (ALEs) or partial dependencies.
We recommend ALEs (see Molnar, 2019) as they are the most reliable and do not calculate predictions for unrealistic combinations of features.
If there are no interactions the ALE estimation will be 0. Figures can be further edited using ggplot2.
}
