% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MrProfilePlots.R
\name{mrProfileplot}
\alias{mrProfileplot}
\title{Wrapper to plot mutlti-response model agnostic profile plots (partial
dependences and accumulated local effects).}
\usage{
mrProfileplot(profileData, sdthresh = 0.05)
}
\arguments{
\item{profileData}{A data frame generated from \code{\link[flashlight:light_profile]{flashlight::light_profile()}}.}

\item{'sdthresh'}{A numeric value used to filter responses that are not
changing across the values of the feature (based on standard deviation).}
}
\description{
Wrapper to plot mutlti-response model agnostic profile plots (partial
dependences and accumulated local effects).
}
\details{
The aim of this function is to plot (1) a reduced set of response
variables that are responding to the feature of choice (plot 1) and the
average ALE of partial dependency  for all responses combined (plot 2). When
there are many responses plot 1 makes interpretation easier by focussing on
the responses changing the most with a feature. The feature selected and plot
type must be specified \code{\link[flashlight:light_profile]{flashlight::light_profile()}} function.
}
\examples{
library(tidymodels)
library(flashlight)

data <- MRFcov::Bird.parasites
Y <- data \%>\%
  select(-scale.prop.zos) \%>\%
  select(order(everything()))
X <- data \%>\%
  select(scale.prop.zos)

model_rf <- rand_forest(
  trees = 100, # 100 trees are set for brevity. Aim to start with 1000
  mode = "classification",
  mtry = tune(),
  min_n = tune()
) \%>\%
  set_engine("randomForest")

mrIML_rf <- mrIMLpredicts(
  X = X,
  Y = Y,
  X1 = Y,
  Model = model_rf,
  prop = 0.7,
  k = 5
)

fl <- mrFlashlight(
  mrIML_rf,
  response = "multi"
)

profileData_pd <- light_profile(fl, v = "scale.prop.zos")
mrProfileplot(profileData_pd, sdthresh = 0.05)[[1]]
mrProfileplot(profileData_pd, sdthresh = 0.05)[[2]]
}
