% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vi_biosecurity.R
\name{plot_vi_biosecurity}
\alias{plot_vi_biosecurity}
\title{Plots global importance (i.e. cumulative importance of all features for all response variables) and individual response importance.}
\usage{
plot_vi_biosecurity(
  VI,
  modelPerf,
  Y,
  X,
  groupCov = NULL,
  cutoff = 0.2,
  plot.pca = "no",
  mode = "regression"
)
}
\arguments{
\item{VI}{A \code{dataframe} dataframe generated from mrvip function}

\item{modelPerf}{A \code{dataframe} dataframe generated by modelPerf function}

\item{Y}{A \code{dataframe} response dataset}

\item{X}{A \code{dataframe} feature dataset}

\item{cutoff}{A \code{numeric} determines  mathews correlation coefifcent (mcc) threshold for displaying individual reponse model. Default is 0.5}

\item{mode}{\code{character}'classification' or 'regression' i.e., is the generative model a regression or classification?}
}
\description{
Plots global importance (i.e. cumulative importance of all features for all response variables) and individual response importance.
}
\details{
1st plot:
and individual reponse feature plots (2nd plot). Requires object generated from the R function mrvip as well as feature data (Y).
Variables also need to be grouped to allow for easier interpretation.
}
\examples{
\dontrun{
groupCov <- c(rep ("Host_characteristics", 1),rep("Urbanisation", 3), rep("Vegetation", 2), rep("Urbanisation",1), rep("Spatial", 2), 
rep('Host_relatedness', 6),rep ("Host_characteristics", 1),rep("Vegetation", 2), rep("Urbanisation",1))
plot_vi(VI=VI,  X=fData,Y=FeaturesnoNA, modelPerf=ModelPerf, groupCov, cutoff= 0.5)}
}
