% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfData.R
\docType{data}
\name{gfData}
\alias{gfData}
\title{Regression data from Fitzpatrick et al. 2014}
\format{
A data frame with the following columns:
\describe{
\item{gfData}{Proportion of individuals in each population with each SNP and bioclimatic data
(bio_1, bio_2, bio_7, bio_10, bio_15, bio_18) as well as spatial MEMs.}
}
}
\source{
Fitzpatrick et al. 2014, Ecology Letters, doi: 10.1111/ele.12376
}
\usage{
gfData
}
\description{
Regression data from Fitzpatrick et al. 2014
}
\examples{
\dontrun{
  head(gfData)
}
}
\keyword{datasets}
