% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfData.R
\docType{data}
\name{gfData}
\alias{gfData}
\title{Regression data from Fitzpatrick et al 2014}
\format{
An object of class \code{data.frame} with 31 rows and 666 columns.
}
\source{
Fitzpatrick et al 2014 Ecology Letters doi: 10.1111/ele.12376
\describe{
\item{gfData} Propotion of individuals in each population with each SNP and bioclim data (bio_1, bio_2, bio_7. bio_10, bio_15, bio_18 as well as spatial MEMs)
}
\usage{
gfData
}
\description{
Regression data from Fitzpatrick et al 2014
}
\examples{
\dontrun{
 gfdata
}
}
\keyword{datasets}
