% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfData.R
\docType{data}
\name{gfData}
\alias{gfData}
\title{Regression data from Fitzpatrick et al 2014}
\format{
An object of class \code{data.frame} with 31 rows and 666 columns.
}
\source{

}
\usage{
gfData
}
\description{
Regression data from Fitzpatrick et al 2014
}
\examples{
\dontrun{
 Responsedata
}
}
\keyword{datasets}
