% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcaret.R
\name{mcaret}
\alias{mcaret}
\title{Perform mutlivariate ensemble learning with models available in caret}
\usage{
mcaret(X, Y, grid)
}
\arguments{
\item{X}{is a data.frame with rows as sites or individuals/populations and columns as loci or species or OTUs.}

\item{Y}{is a data.frame with rows as sites or individuals/populations and columns as features e.g. environmental or host or demographic variables.}

\item{grid}{is the model specific grid of values to tune each model e.g. random forest. Not quite working properly yet}
}
\description{
This function allows you to perform ensemble learning on mutliple response variables
}
\keyword{ensemble}
\keyword{learning}
\keyword{multivariate}
