% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MrFlashLight.R
\name{mrFlashlight}
\alias{mrFlashlight}
\title{mrFlashlight:  Wrapper to plot mutlti-response model agnostic partial dependency plots.}
\usage{
mrFlashlight(yhats, X, Y, response = "multi", index = 1)
}
\arguments{
\item{yhats}{A \code{list} is the list generated by mrIMLpredicts}

\item{X}{A \code{dataframe} is a response variable data set (species, OTUs, SNPs etc)}

\item{Y}{A \code{dataframe} #is the feature dataset}

\item{response}{\code{character} 'single' selects one response, 'multi' selects all responses}

\item{index}{\code{numeric} selects which response to create a flaslight object for. Only used when 'single' is selcted. The order is the same as 'X'.}
}
\description{
mrFlashlight:  Wrapper to plot mutlti-response model agnostic partial dependency plots.
}
\details{
The aim of this function is to enable users to ustilize interpretable machine learning methods
to understand their multi-response and single response models
}
