% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MrFlashLight.R
\name{mrFlashlight}
\alias{mrFlashlight}
\title{mrFlashlight:  Wrapper to run multi-response model agnostic interpretable machine learning analyses.}
\usage{
mrFlashlight(
  yhats,
  X,
  X1,
  Y,
  response = "multi",
  index = 1,
  mode = "regression",
  predict_function = NULL
)
}
\arguments{
\item{yhats}{A \code{list} is the list generated by mrIMLpredicts}

\item{X}{A \code{dataframe} is the feature data set}

\item{Y}{A \code{dataframe} is a response variable data set (species, OTUs, SNPs etc)}

\item{response}{\code{character} single' selects one response, 'multi' selects all responses}

\item{index}{\code{numeric} selects which response to create a flashlight object for. Only used when 'single' is selected. The order is the same as 'Y'.}

\item{mode}{\code{character}'classification' or 'regression' i.e., is the generative model a regression or classification?}

\item{predict_function}{\code{function}  user specified predict function}
}
\description{
mrFlashlight:  Wrapper to run multi-response model agnostic interpretable machine learning analyses.
}
\details{
The aim of this function is to enable users to utilize interpretable machine learning methods
to understand their multi-response and single response models
}
\examples{
#single response
fl <- mrFlashlight(yhats, X, Y, response = "single", index=1, mode = "regression")

plot(light_performance(fl), fill = "orange") + labs(x = element_blank())

plot(light_breakdown(fl , new_obs = cbind(X, Y)[1, ]),by = X, v=Y) #prints all responses - need to fix but could be quite handy.

int <- light_interaction(fl, pairwise=TRUE) #not working, but possible!

Multiple response
flashlightObj <- mrFlashlight(yhats, X, Y, response = "multi", mode='regression')
}
