% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MrFlashLight.R
\name{mrFlashlight}
\alias{mrFlashlight}
\title{mrFlashlight: Wrapper to run multi-response model agnostic interpretable machine learning analyses.}
\usage{
mrFlashlight(
  yhats,
  X,
  X1,
  Y,
  response = "multi",
  index = 1,
  mode = "regression",
  predict_function = NULL
)
}
\arguments{
\item{yhats}{A \code{list} generated by \code{mrIMLpredicts}.}

\item{X}{A \code{data.frame} containing the feature data set.}

\item{X1}{A \code{data.frame} of additional predictors (optional).}

\item{Y}{A \code{data.frame} containing the response variable data set (species, OTUs, SNPs, etc.).}

\item{response}{\code{character} indicating the type of response: 'single' selects one response, 'multi' selects all responses.}

\item{index}{\code{numeric} used when \code{response} is 'single' to select which response to create a flashlight object for. The order is the same as in \code{Y}.}

\item{mode}{\code{character} indicating the type of model: 'classification' or 'regression'.}

\item{predict_function}{\code{function} specifying a user-defined prediction function (optional).}
}
\description{
mrFlashlight: Wrapper to run multi-response model agnostic interpretable machine learning analyses.
}
\details{
This function enables users to utilize interpretable machine learning methods to understand their multi-response and single-response models.
}
\examples{
# Single response
fl <- mrFlashlight(yhats, X, Y, response = "single", index = 1, mode = "regression")
plot(light_performance(fl), fill = "orange") + labs(x = element_blank())
plot(light_breakdown(fl, new_obs = cbind(X, Y)[1, ]), by = X, v = Y)
int <- light_interaction(fl, pairwise = TRUE)

# Multiple response
flashlightObj <- mrFlashlight(yhats, X, Y, response = "multi", mode = "regression")
}
