% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrIMLperformance.R
\name{mrIMLperformance}
\alias{mrIMLperformance}
\title{Wrapper to calculate performance metrics (Mathews correlation coefficient, sensitivity and specificity) for each model for each response variable.}
\usage{
mrIMLperformance(yhats, Model, Y, mode = "regression")
}
\arguments{
\item{yhats}{A \code{list} is the list generated by mrIMLpredicts}

\item{Model}{A \code{list}  the model used to generate the yhats object}

\item{Y}{A \code{dataframe} is a response variable data set (species, SNPs etc).}

\item{mode}{\code{character}'classification' or 'regression' i.e., is the generative model a regression or classification?}
}
\description{
Wrapper to calculate performance metrics (Mathews correlation coefficient, sensitivity and specificity) for each model for each response variable.
}
\details{
Outputs a dataframe of commonly used metric that can be used to compare model performance of classification models. Performance metrics are based on testing data. But MCC is useful (higher numbers = better fit)
}
