% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrIMLperformance.R
\name{mrIMLperformance}
\alias{mrIMLperformance}
\title{Wrapper to calculate performance metrics (Mathews correlation coefficent, sensitivity and specificity) for each model for each response variable.}
\usage{
mrIMLperformance(yhats, model1, X)
}
\arguments{
\item{yhats}{A \code{list} is the list generated by MrTidyPerf}

\item{model1}{A \code{list}  #the model used to generate the yhats object}

\item{X}{A \code{dataframe} #is a response variable data set (specie, SNPs etc).}
}
\description{
Wrapper to calculate performance metrics (Mathews correlation coefficent, sensitivity and specificity) for each model for each response variable.
}
\details{
Outputs a dataframe of commonly used metric that can be used to compare model
performance of classification models. Performance metrics are based on testing data. But MCC is useful (higher numbers = better fit)
}
