% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrIMLpredicts.R
\name{mrIMLpredicts}
\alias{mrIMLpredicts}
\title{mrIMLpredicts:  Wrapper to generate multi-response tidy models.This function produces yhats that are used in the stack function as well 
as saving all model characteristics for subsequent functions.}
\usage{
mrIMLpredicts(X, Y, model1, balance_data = "no", calculate.resid = "yes")
}
\arguments{
\item{X}{A \code{dataframe} represents predictor or feature data.}

\item{Y}{A \code{dataframe} is a response variable data set (species, OTUs, SNPs etc).}

\item{balance_data}{A \code{character} 'up', 'down' or 'no'.}

\item{calculate.resid}{A \code{character} 'yes',or 'no'. We can't calculate deviance residuals for multinomial models
currently.}

\item{Model}{1 A \code{list} can be any model from the tidy model package. See examples.}
}
\description{
This function fits separate classication models for each response variable in a dataset.
}
\details{
Y (response variables) should be binary (0/1). Rows in X (features) have the same id (host/site/population)
 as Y. 
 Class imblanace can be a real issue for classification analyses. Class imbalance can be addressed for each
  response variable using 'up' (upsampling using ROSE bootstrapping), 'down' (downsampling) 
or 'no' (no balancing of classes).
Calculate deviance residuals
}
\examples{
model1 <- #model used to generate yhat
specify that the model is a random forest
logistic_reg() \%>\%
# select the engine/package that underlies the model
set_engine("glm") \%>\%
 # choose either the continuous regression or binary classification mode
 set_mode("classification")
 
}
