% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrIMLpredicts.R
\name{mrIMLpredicts}
\alias{mrIMLpredicts}
\title{Wrapper to generate multi-response predictive models.}
\usage{
mrIMLpredicts(
  X,
  Y,
  Model,
  balance_data = "no",
  mode = "regression",
  transformY = "log",
  dummy = FALSE,
  tune_grid_size = 10,
  k = 10,
  seed = sample.int(1e+08, 1)
)
}
\arguments{
\item{X}{A \code{dataframe} represents predictor or feature data.}

\item{Y}{A \code{dataframe} is a response variable data (species, OTUs, SNPs etc).}

\item{Model}{1 A \code{list} can be any model from the tidy model package. See examples.}

\item{balance_data}{A \code{character} 'up', 'down' or 'no'.}

\item{mode}{\code{character}'classification' or 'regression' i.e., is the generative model a regression or classification?}

\item{dummy}{A \code{logical} 'TRUE or FALSE'.}

\item{tune_grid_size}{A \code{numeric} sets the grid size for hyperparamter tuning. Larger grid sizes increase computational time.}

\item{k}{A \code{numeric} sets the number of folds in the 10-fold cross-validation. 10 is the default.}
}
\description{
Wrapper to generate multi-response predictive models.
}
\details{
This function produces yhats that used in all subsequent functions.
This function fits separate classification/regression models for each response variable in a data set.  Rows in X (features) have the same id (host/site/population)
as Y. Class imbalance can be a real issue for classification analyses. Class imbalance can be addressed for each
response variable using 'up' (upsampling using ROSE bootstrapping), 'down' (downsampling)
or 'no' (no balancing of classes).
}
