% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrIMLpredicts.R
\name{mrIMLpredicts}
\alias{mrIMLpredicts}
\title{Wrapper to generate multi-response predictive models.}
\usage{
mrIMLpredicts(
  X,
  Y,
  model1,
  balance_data = "no",
  model = "regression",
  parallel = TRUE,
  transformY = "log",
  tune_grid_size = 10,
  seed = sample.int(1e+08, 1)
)
}
\arguments{
\item{X}{A \code{dataframe} represents predictor or feature data.}

\item{Y}{A \code{dataframe} is a response variable data set (species, OTUs, SNPs etc).}

\item{balance_data}{A \code{character} 'up', 'down' or 'no'.}

\item{tune_grid_size}{A \code{numeric} sets the grid size for hyperparamter tuning. Larger grid sizes increase computational time.}

\item{Model}{1 A \code{list} can be any model from the tidy model package. See examples.}
}
\description{
Wrapper to generate multi-response predictive models.
}
\details{
This function produces yhats that used in all model characteristics for subsequent functions.
This function fits separate classication models for each response variable in a dataset. Y (response variables) should be binary (0/1). Rows in X (features) have the same id (host/site/population)
as Y. Class imblanace can be a real issue for classification analyses. Class imbalance can be addressed for each
response variable using 'up' (upsampling using ROSE bootstrapping), 'down' (downsampling)
or 'no' (no balancing of classes).
}
