% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrInteractions.R
\name{mrInteractions}
\alias{mrInteractions}
\title{mrInteractions:  Wrapper to calculate and plot summed interactions of features across response variables. Bases on Greenwell et al 2018}
\usage{
mrInteractions(yhats, X, Y)
}
\arguments{
\item{yhats}{A \code{list} is the list generated by mrIMLpredicts}

\item{X}{A \code{dataframe} is a response variable data set (species, OTUs, SNPs etc)}

\item{Y}{A \code{dataframe} #is the feature dataset}
}
\description{
mrInteractions:  Wrapper to calculate and plot summed interactions of features across response variables. Bases on Greenwell et al 2018
}
\details{
The aim of this function is to enable users to calculate interaction importance for all fatures in a data set.
and then use plot_mrpd (a plotting function) to create a pd plot for all species/SNPs for each feature. 
Could build functionality for ICE plots but that would be messy for multiple repsonse variables.
}
