% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrLocalExplainer.R
\name{mrLocalExplainer}
\alias{mrLocalExplainer}
\title{Run local explanation methods for individual data points}
\usage{
mrLocalExplainer(my.data, model_data, outcome)
}
\arguments{
\item{my.data}{A \code{dataframe} data frame of predictor variables}

\item{model_data}{A \code{workflow} workflow object containing the machine learning model}

\item{outcome}{\code{vector} vector containing the outcome for each data instance}
}
\value{

}
\description{
Run local explanation methods for individual data points
}
\examples{
#mrLocalExplainer(data, yhats, data$Class)
## will return matrix of phi values, individualized and aggregated plots 

}
