% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrPdPlotBootstrap.R
\name{mrPdPlotBootstrap}
\alias{mrPdPlotBootstrap}
\title{Bootstrap Partial Dependence plots}
\usage{
mrPdPlotBootstrap(
  mrIML_obj,
  mrBootstrap_obj,
  vi_obj = NULL,
  target,
  global_top_var = 2
)
}
\arguments{
\item{mrIML_obj}{A list object returned by \code{\link[=mrIMLpredicts]{mrIMLpredicts()}}.}

\item{mrBootstrap_obj}{A list object returned by \code{\link[=mrBootstrap]{mrBootstrap()}}.}

\item{vi_obj}{A list object returned by \code{\link[=mrVip]{mrVip()}}. If \code{vi_obj} is not
provided then it is created inside \code{mrPD_bootstrap} by running \code{\link[=mrVip]{mrVip()}}}

\item{target}{The target variable for generating plots.}

\item{global_top_var}{The number of top variables to consider (default: 2).}
}
\value{
A list with two elements:
*\verb{[[1]]} A data frame of the partial dependence grid for each response model, predictor variable, and bootstrap
*\verb{[[2]]} A list of partial dependence plots for each predictor variable in the \code{target} response model.
}
\description{
This function bootstraps model predictions and generates partial dependence
plots for each response variable. It also creates a combined plot for the top
variables of interest.
}
\examples{
library(tidymodels)

data <- MRFcov::Bird.parasites
Y <- data \%>\%
  select(-scale.prop.zos) \%>\%
  dplyr::select(order(everything()))
X <- data \%>\%
  select(scale.prop.zos)

model_rf <- rand_forest(
  trees = 100, # 100 trees are set for brevity. Aim to start with 1000
  mode = "classification",
  mtry = tune(),
  min_n = tune()
) \%>\%
  set_engine("randomForest")

mrIML_rf <- mrIMLpredicts(
  X = X,
  Y = Y,
  X1 = Y,
  Model = model_rf,
  prop = 0.7,
  k = 5
)

mrIML_rf_boot <- mrIML_rf \%>\%
  mrBootstrap()
  
mrIML_rf_PD <- mrPdPlotBootstrap(
  mrIML_rf,
  mrIML_rf_boot,
  target = "Plas",
  global_top_var = 4
)

head(mrIML_rf_PD[[1]])
mrIML_rf_PD[[2]]

}
