% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrPerformancePlot.R
\name{mrPerformancePlot}
\alias{mrPerformancePlot}
\title{Multiple Response Variable Selection}
\usage{
mrPerformancePlot(
  ModelPerf1 = NULL,
  ModelPerf2 = NULL,
  mod_names = c("combined", "Xonly_model"),
  mode = "classification"
)
}
\arguments{
\item{X}{Dataframe of predictors. (Optional)}

\item{X1}{Dataframe of additional predictors for predicting the response variables. (Optional)}

\item{Y}{Dataframe with response variables, where column names are the response variable names.}
}
\value{
A list where each element corresponds to a response variable, containing the Boruta result object for that response variable.
}
\description{
Identify predictors not useful in the prediction of multiple response variables using the Boruta algorithm.
}
\examples{
X <- data.frame(matrix(rnorm(1000), ncol = 10))
X1 <- data.frame(matrix(rnorm(100), ncol = 5))
Y <- data.frame(response1 = rnorm(100), response2 = rnorm(100))
results <- mrVariableSelect(X, X1, Y)

}
