% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrPlotInteractions.R
\name{mrPlot_interactions}
\alias{mrPlot_interactions}
\title{Plots global interactions as well as individual response interaction importance.}
\usage{
mrPlot_interactions(interactions, X, Y, top_ranking = 3, top_response = 10)
}
\arguments{
\item{interactions}{A \code{dataframe} data frame generated from mrInteractioms function}

\item{X}{A \code{dataframe} feature data set}

\item{Y}{A \code{dataframe} response data set}

\item{top_ranking}{A \code{numeric} determines how many of the strongest feature interacions to view/include}

\item{top_response}{A \code{numeric} how many of the response variables with the strongest interactions to view}
}
\description{
Plots global interactions as well as individual response interaction importance.
}
\details{
1st plot: Barplots showing the mean and cumulative importance of each of the top pairs of interactions in the model.
2nd plot: Barplot of the responses with the strongest interactions
3rd plot: Barplots of the strongest interactions for each of the top response variables.
interactions <-mrInteractions(yhats, X, Y) #this is computationally intensive so multicores are needed. If stopped prematurely - have to reload things
mrPlot_interactions(Interact, X,Y, top_ranking = 3, top_response=3)
}
