% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MrProfilePlots.R
\name{mrProfileplot}
\alias{mrProfileplot}
\title{mrProfileplot:  Wrapper to plot mutlti-response model agnostic profile plots (partial dependences and accumulated local effects).}
\usage{
mrProfileplot(profileData, sdthresh = 0.05)
}
\arguments{
\item{profileData}{\code{dataframe} is the data generated from the flashlight packages 'light_profile' function.}

\item{'sdthresh'}{\code{numeric} value used to filter responses that are not changing across the values of the feature (based on standard deviation)}
}
\description{
mrProfileplot:  Wrapper to plot mutlti-response model agnostic profile plots (partial dependences and accumulated local effects).
}
\details{
The aim of this function is to plot (1) a reduced set of response variables that are responding to the feature of choice (plot 1)
and the average ALE of partial dependency  for all responses combined (plot 2). When there are many responses plot 1 makes interpretation easier by focussing on the
the responses changing the most with a feature. The feature selected and plot type must be specified 'light_profile' function
}
