% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrvip.R
\name{mrVip}
\alias{mrVip}
\title{Wrapper to estimate model-agnostic variable importance for multi-response models.}
\usage{
mrVip(yhats, X, Y)
}
\arguments{
\item{yhats}{A \code{list} is the list generated by mrIMLpredicts}

\item{X}{A \code{dataframe} is the feature data}

\item{Y}{A \code{dataframe} is the response data}
}
\value{
A dataframe containing variable importance scores for each response variable
}
\description{
Wrapper to estimate model-agnostic variable importance for multi-response models.
}
\details{
Calculates variable importance using based on partial dependencies but could do permutations as well (more memory intensive).
Key input is the object created by MrIML. Can be plotted with the plot_vi function.
}
\examples{
VI <- mrVip(yhats, X=X, Y=Y)
groupCov <- c(rep ("Host_characteristics", 1),rep("Urbanisation", 3), rep("Vegetation", 2), rep("Urbanisation",1), rep("Spatial", 2), 
rep('Host_relatedness', 6),rep ("Host_characteristics", 1),rep("Vegetation", 2), rep("Urbanisation",1))  
plot_vi(VI=VI,  X=X,Y=Y, modelPerf=ModelPerf, groupCov, cutoff= 0.5)

}
