% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrvip.R
\name{mrVip}
\alias{mrVip}
\title{Wrapper to estimate model-agnostic variable importance for multi-response models.}
\usage{
mrVip(yhats, Y)
}
\arguments{
\item{yhats}{A \code{list} is the list generated by mrIMLpredicts}

\item{Y}{A \code{dataframe} #is the feature dataset}
}
\description{
Wrapper to estimate model-agnostic variable importance for multi-response models.
}
\details{
Calculates variabie importance using based on partial dependencies but could do permutation as well (mor memory intensive).
Key imput is the object created by MrIML. Can be plotted with the plot_vi function.
}
