% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrvip.R, R/mrvip_v2.R
\name{mrVip}
\alias{mrVip}
\title{Wrapper to estimate model-agnostic variable importance for multi-response models.}
\usage{
mrVip(yhats, X)

mrVip(yhats, X)
}
\arguments{
\item{yhats}{A \code{list} is the list generated by mrIMLpredicts}

\item{X}{A \code{dataframe} is the feature data}

\item{Y}{A \code{dataframe} #is the feature data set}
}
\description{
Wrapper to estimate model-agnostic variable importance for multi-response models.

Wrapper to estimate model-agnostic variable importance for multi-response models.
}
\details{
Calculates variable importance using based on partial dependencies but could do permutations as well (more memory intensive).
Key input is the object created by MrIML. Can be plotted with the plot_vi function.

Calculates variabie importance using based on partial dependencies but could do permutation as well (mor memory intensive).
Key imput is the object created by MrIML. Can be plotted with the plot_vi function.
}
