% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrvip.R
\name{mrvip}
\alias{mrvip}
\title{Calculates and helps interpret variable importance for mrIML models.}
\usage{
mrvip(
  yhats = NULL,
  mrBootstrap_obj = NULL,
  X = X,
  X1 = NULL,
  Y = Y,
  mode = NULL,
  threshold = 0.1,
  global_top_var = 10,
  local_top_var = 5,
  taxa = NULL,
  ModelPerf = ModelPerf
)
}
\arguments{
\item{yhats}{A list of model predictions.}

\item{mrBootstrap_obj}{The object containing model bootstrapping results.}

\item{X}{The predictor data.}

\item{X1}{A \code{dataframe} extra predictor set used in each model. For the MrIML Joint species distribution model (JSDM) this is just a copy of the response data.}

\item{Y}{The response data.}

\item{mode}{\code{character}'classification' or 'regression' i.e., is the generative model a regression or classification?}

\item{threshold}{The threshold for model performance (AUC) below which variables are filtered (default: 0.1).}

\item{global_top_var}{The number of top global variables to display (default: 10).}

\item{local_top_var}{The number of top local variables for each response to display (default: 5).}

\item{ModelPerf}{A list containing model performance metrics for each response variable.}
}
\value{
A list containing variable importance data and a combined plot.
}
\description{
Calculates and helps interpret variable importance for mrIML models.
}
