% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrVip.R
\name{mrVipPCA}
\alias{mrVipPCA}
\title{Principle Component Analysis of mrIML variable importance}
\usage{
mrVipPCA(mrVip_obj)
}
\arguments{
\item{mrVip_obj}{A list returned by \code{\link[=mrVip]{mrVip()}}.}
}
\value{
A list of PCA results:
\itemize{
\item $PCA_plot
\item $PC_outlires
\item $eigenvalues
\item $PC_scores
}
}
\description{
Principle Component Analysis of mrIML variable importance
}
