% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrvip.R
\name{mrVipPCA}
\alias{mrVipPCA}
\title{Principle Component Analysis of \code{mrIML} variable importance}
\usage{
mrVipPCA(mrVip_obj)
}
\arguments{
\item{mrVip_obj}{A list returned by \code{\link[=mrvip]{mrvip()}}.}
}
\value{
A list of PCA results:
\itemize{
\item \verb{$PCA_plot}
\item \verb{$PC_outlires}
\item \verb{$eigenvalues}
\item \verb{$PC_scores}
}
}
\description{
Principle Component Analysis of \code{mrIML} variable importance
}
