% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSnpsPed.R
\name{readSnpsPed}
\alias{readSnpsPed}
\title{Conversion to single column per locus from plink file via LEA functionality}
\usage{
readSnpsPed(pedfile, mapfile)
}
\arguments{
\item{pedfile}{A \code{dataframe} dataframe generated from mrvip function}

\item{mapfile}{A \code{dataframe} response dataset}
}
\description{
Conversion to single column per locus from plink file via LEA functionality
}
\details{
Function to import SNP data from a plink format into a format suitable for MrIML predicts
(presence/absence of an alelle for each locus).
Currently if there is missing data (NAs) it either imputes them as the mode or leaves them. A histogram is also produced of the missing data.
}
\examples{
\dontrun{snps <- readSnpsPed("bobcat.plink.ped", "bobcat.plink.map.map") #NAs in data and interpolated as the mode. 
X <- filterRareCommon (snps, lower=0.4, higher=0.7) #these are harsh}
}
