% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResistanceComponents.R
\name{resist_components}
\alias{resist_components}
\title{Resist_components :  Calculates resistance components from a list of pairwise resistance surfaces.}
\usage{
resist_components(filename = filename, p_val = p_val)
}
\arguments{
\item{p_val}{A \code{numeric} this sets the significance threshold for axes in explaining variance in the original resistance matrix based on redundancy analysis. In effect this filters out axes that dont explain variance.}

\item{fileneame}{A \code{character} this is the location where the resistance surfaces are stored.}
}
\description{
Resist_components :  Calculates resistance components from a list of pairwise resistance surfaces.
}
\details{
Outputs a data frame of significant resistance components for each matrix in the target folder. These data can be combined with non-pairwise matrix data.
}
